+++ apps/browser/src/autofill/fido2/background/abstractions/fido2.background.ts
