+++ apps/browser/src/autofill/fido2/background/fido2.background.spec.ts
@@ -18,7 +18,7 @@ import {
 } from "../../../autofill/spec/testing-utils";
 import { BrowserApi } from "../../../platform/browser/browser-api";
 import { BrowserScriptInjectorService } from "../../../platform/services/browser-script-injector.service";
-import { AbortManager } from "../../background/abort-manager";
+import { AbortManager } from "../../../vault/background/abort-manager";
 import { Fido2ContentScript, Fido2ContentScriptId } from "../enums/fido2-content-script.enum";
 import { Fido2PortName } from "../enums/fido2-port-name.enum";
