+++ apps/browser/src/autofill/fido2/background/fido2.background.ts
@@ -13,7 +13,7 @@ import { VaultSettingsService } from "@bitwarden/common/vault/abstractions/vault
 
 import { BrowserApi } from "../../../platform/browser/browser-api";
 import { ScriptInjectorService } from "../../../platform/services/abstractions/script-injector.service";
-import { AbortManager } from "../../background/abort-manager";
+import { AbortManager } from "../../../vault/background/abort-manager";
 import { Fido2ContentScript, Fido2ContentScriptId } from "../enums/fido2-content-script.enum";
 import { Fido2PortName } from "../enums/fido2-port-name.enum";
