+++ apps/browser/src/autofill/fido2/content/fido2-content-script.ts
