+++ apps/browser/src/autofill/fido2/content/fido2-page-script-append.mv2.ts
