+++ apps/browser/src/autofill/fido2/content/fido2-page-script.webauthn-supported.spec.ts
@@ -5,7 +5,7 @@ import {
   createCredentialRequestOptionsMock,
   setupMockedWebAuthnSupport,
 } from "../../../autofill/spec/fido2-testing-utils";
-import { WebauthnUtils } from "../webauthn-utils";
+import { WebauthnUtils } from "../../../vault/fido2/webauthn-utils";
 
 import { MessageType } from "./messaging/message";
 import { Messenger } from "./messaging/messenger";
@@ -40,7 +40,7 @@ jest.mock("./messaging/messenger", () => {
     },
   };
 });
-jest.mock("../webauthn-utils");
+jest.mock("../../../vault/fido2/webauthn-utils");
 
 describe("Fido2 page script with native WebAuthn support", () => {
   (jest.spyOn(globalThis, "document", "get") as jest.Mock).mockImplementation(
@@ -54,7 +54,7 @@ describe("Fido2 page script with native WebAuthn support", () => {
   setupMockedWebAuthnSupport();
 
   beforeAll(() => {
-    require("./page-script");
+    require("./fido2-page-script");
   });
 
   afterEach(() => {
@@ -154,7 +154,7 @@ describe("Fido2 page script with native WebAuthn support", () => {
         contentType: "json/application",
       }));
 
-      require("./content-script");
+      require("./fido2-content-script");
 
       expect(Messenger.forDOMCommunication).not.toHaveBeenCalled();
     });
@@ -172,7 +172,7 @@ describe("Fido2 page script with native WebAuthn support", () => {
         },
       }));
 
-      require("./content-script");
+      require("./fido2-content-script");
 
       expect(Messenger.forDOMCommunication).not.toHaveBeenCalled();
     });
