+++ apps/browser/src/autofill/fido2/content/fido2-page-script.webauthn-unsupported.spec.ts
@@ -4,7 +4,7 @@ import {
   createCredentialCreationOptionsMock,
   createCredentialRequestOptionsMock,
 } from "../../../autofill/spec/fido2-testing-utils";
-import { WebauthnUtils } from "../webauthn-utils";
+import { WebauthnUtils } from "../../../vault/fido2/webauthn-utils";
 
 import { MessageType } from "./messaging/message";
 import { Messenger } from "./messaging/messenger";
@@ -39,7 +39,7 @@ jest.mock("./messaging/messenger", () => {
     },
   };
 });
-jest.mock("../webauthn-utils");
+jest.mock("../../../vault/fido2/webauthn-utils");
 
 describe("Fido2 page script without native WebAuthn support", () => {
   (jest.spyOn(globalThis, "document", "get") as jest.Mock).mockImplementation(
@@ -50,7 +50,7 @@ describe("Fido2 page script without native WebAuthn support", () => {
   const mockCreateCredentialsResult = createCreateCredentialResultMock();
   const mockCredentialRequestOptions = createCredentialRequestOptionsMock();
   const mockCredentialAssertResult = createAssertCredentialResultMock();
-  require("./page-script");
+  require("./fido2-page-script");
 
   afterEach(() => {
     jest.resetModules();
