+++ apps/browser/src/manifest.json
@@ -101,7 +101,7 @@
     }
   },
   "web_accessible_resources": [
-    "content/fido2/page-script.js",
+    "content/fido2-page-script.js",
     "content/lp-suppress-import-download.js",
     "notification/bar.html",
     "images/icon38.png",
