+++ apps/browser/src/manifest.v3.json
@@ -107,7 +107,7 @@
   "web_accessible_resources": [
     {
       "resources": [
-        "content/fido2/page-script.js",
+        "content/fido2-page-script.js",
         "notification/bar.html",
         "images/icon38.png",
         "images/icon38_locked.png",
